Renamed fake_payment to payment, fixed some formatting in customer bill
