+++ templates/customer/customer_bill.html
@@ -40,7 +40,7 @@
                             <th class="text-center">Qty</th>
                             <th class="text-end">Price</th>
                             <th class="text-end">Total</th>
-                            <th>Assigned To</th>
+                            <th class="text-center">Assigned To</th>
                             <th class="text-center">Action</th>
                         </tr>
                     </thead>
@@ -51,7 +51,7 @@
                             <td class="text-center">{{ item.quantity }}</td>
                             <td class="text-end">${{ "%.2f"|format(item.price) }}</td>
                             <td class="text-end fw-bold">${{ "%.2f"|format(item.price * item.quantity) }}</td>
-                            <td>
+                            <td class="text-center">
                                 {% if item.assigned_to and item.assigned_to|length > 0 %}
                                 <div class="d-flex flex-wrap gap-1">
                                     {% for user_id in item.assigned_to %}
@@ -70,7 +70,7 @@
                                 <!-- Assign to current user button -->
                                 {% if not item.assigned_to or current_user.id not in item.assigned_to %}
                                 <form method="POST"
-                                    action="{{ url_for('customer.split_bill', group_id=group._id, item_id=item._id, user_id=current_user.id) }}"
+                                    action="{{ url_for('customer.display_bill', group_id=group._id, item_id=item._id, user_id=current_user.id) }}"
                                     style="display:inline;">
                                     <button type="submit" class="btn btn-sm btn-success" title="Assign to yourself">
                                         <i class="fa fa-plus"></i>
@@ -101,10 +101,10 @@
             <strong>Bill Subtotal:</strong> ${{ "%.2f"|format(bill.subtotal) }}
         </p>
         <p>
-            <strong>Tax ({{ "%.2f"|format(tax) }}%):</strong> ${{ "%.2f"|format(bill.subtotal * tax) }}
+            <strong>Tax ({{ "%.2f"|format(tax) }}%):</strong> ${{ "%.2f"|format(bill.subtotal * tax / 100) }}
         </p>
         <p class="fw-bold text-success">
-            <strong>Bill Total:</strong> ${{ "%.2f"|format(bill.subtotal * (1 + tax)) }}
+            <strong>Bill Total:</strong> ${{ "%.2f"|format(bill.subtotal * (1 + tax / 100)) }}
         </p>
     </div>
